+++ packages/openfort-react/src/components/ConnectModal/ConnectUsing.tsx
@@ -64,6 +64,7 @@ const ConnectUsing = () => {
             switchConnectMethod={(id?: string) => {
               //if (id) setId(id);
               setStatus(states.INJECTOR);
+              setTimeout(context.triggerResize, 10); // delay required here for modal to resize
             }}
           />
         </motion.div>
@@ -80,6 +81,7 @@ const ConnectUsing = () => {
             switchConnectMethod={(id?: string) => {
               //if (id) setId(id);
               setStatus(states.QRCODE);
+              setTimeout(context.triggerResize, 10); // delay required here for modal to resize
             }}
           />
         </motion.div>