+++ packages/openfort-react/src/utils/index.ts
@@ -34,7 +34,7 @@ const nFormatter = (num: number, digits: number = 2) => {
     { value: 1, symbol: '' },
     { value: 1e3, symbol: 'k' },
     { value: 1e6, symbol: 'm' },
-    { value: 1e9, symbol: 'g' },
+    { value: 1e9, symbol: 'b' },
     { value: 1e12, symbol: 't' },
     { value: 1e15, symbol: 'p' },
     { value: 1e18, symbol: 'e' },
@@ -90,6 +90,12 @@ function flattenChildren(children: React.ReactNode): ReactChildArray {
 export const isWalletConnectConnector = (connectorId?: string) =>
   connectorId === 'walletConnect';
 
+export const isFamilyAccountsConnector = (connectorId?: string) =>
+  connectorId === 'familyAccountsProvider';
+
+export const isFamilyConnector = (connectorId?: string) =>
+  connectorId === 'co.family.wallet';
+
 export const isMetaMaskConnector = (connectorId?: string) =>
   connectorId === 'metaMaskSDK';
 
@@ -98,6 +104,9 @@ export const isCoinbaseWalletConnector = (connectorId?: string) =>
 
 export const isLedgerConnector = (connectorId?: string) =>
   connectorId === 'ledger';
+
+export const isPortoConnector = (connectorId?: string) =>
+  connectorId === 'xyz.ithaca.porto';
 
 export const isSafeConnector = (connectorId?: string) => connectorId === 'safe';